memory_service/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Storage backends
asyncpg>=0.29.0
pgvector>=0.2.4
neo4j>=5.16.0
redis[hiredis]>=5.0.0

# ML/Embeddings
numpy>=1.26.0
openai>=1.0.0  # Optional, for OpenAI embeddings
# sentence-transformers>=2.2.0  # Alternative for local embeddings

# Utilities
ulid-py>=1.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.44b0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
ruff>=0.1.0
mypy>=1.8.0
